fix 修复CourseBo的学分为Double类型

+++ bluett-ssms/src/main/java/org/bluett/ssms/domain/bo/CourseBo.java
@@ -2,14 +2,14 @@ package org.bluett.ssms.domain.bo;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import javax.validation.constraints.*;
-
-import java.util.Date;
-import com.fasterxml.jackson.annotation.JsonFormat;
 import org.bluett.common.core.domain.BaseEntity;
 import org.bluett.common.core.validate.AddGroup;
 import org.bluett.common.core.validate.EditGroup;
 import org.springframework.format.annotation.DateTimeFormat;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import java.util.Date;
 
 /**
  * 课程信息业务对象 ssms_course
@@ -44,7 +44,7 @@ public class CourseBo extends BaseEntity {
      * 学分
      */
     @NotNull(message = "学分不能为空", groups = { AddGroup.class, EditGroup.class })
-    private Long credit;
+    private Double credit;
 
     /**
      * 开始日期
